SpreadsheetUI : Split code into separate module
`SpreadsheetUI.py` had become somewhat sizeable. The up-coming refactor
to a shared model approach would cause this to grow further with the
addition of the various proxy models. This commit should have no outward
behaviour or API changes. It simply splits out individual classes into
their own files to aid maintenance.

Aside from minor changes required to avoid circular imports, code structure
has been left as-is. Some floating functions have been moved into the
consuming class where possible.

There is probably some future work rationalising what is in `_Algo`, vs
private methods in `_Menus` or other class implementations, but this
is beyond the scope of this initial refactor.
